requirements.txt
alembic==0.9.9
amqp==2.2.2
asgi-redis==1.4.3
asgiref==2.1.1
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.5.1
Automat==0.6.0
Babel==2.5.3
backcall==0.1.0
billiard==3.5.0.3
bleach==2.1.3
celery==4.1.0
certifi==2018.1.18
cffi==1.11.5
channels==2.1.1
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.2
daphne==2.1.1
decorator==4.2.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.4
django-allauth==0.35.0
django-bower==5.2.0
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-configurations==2.0
django-cors-headers==2.2.0
django-extensions==2.0.7
django-filter==1.1.0
django-polybuilder==0.2.2
django-rest-auth==0.9.3
djangorestframework==3.8.2
entrypoints==0.2.3
eventlet==0.22.1
flower==0.9.2
gevent==1.2.2
greenlet==0.4.13
h2==3.0.1
hpack==3.0.0
html5lib==1.0.1
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.0
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterhub==0.8.1
jupyterlab==0.31.12
jupyterlab-launcher==0.10.5
kombu==4.1.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
mistune==0.8.3
msgpack-python==0.5.6
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.6
nose==1.3.7
notebook==4.4.1
oauthlib==2.0.7
pamela==0.3.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pgcontents==0.5.3
pickleshare==0.7.4
pkg-resources==0.0.0
pluggy==0.6.0
priority==1.3.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
py==1.5.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==17.5.0
pyserial==3.4
python-dateutil==2.7.2
python-editor==1.0.3
python-engineio==2.0.4
python-oauth2==1.1.0
python-socketio==1.9.0
python3-openid==3.1.0
pytz==2018.4
pyusb==1.0.2
PyVISA==1.9.0
PyVISA-sim==0.3
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.6
stringparser==0.4.1
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
tox==2.9.1
traitlets==4.3.2
Twisted==18.4.0
txaio==2.10.0
urllib3==1.22
vine==1.1.4
virtualenv==15.2.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.2.0
zope.interface==4.5.0
